+++ src/domain/user-accounts-actions/user-actions.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { UserAccountsActionsController } from './controller/account-actions.controller';
-import { AccountsModule } from 'src/accounts/accounts.module';
-import { RecordsModule } from 'src/records/records.module';
-import { BudgetsModule } from 'src/budgets/budgets.module';
-import { BudgetHistoryModule } from 'src/budget-history/budget-history.module';
+import { AccountsModule } from '../../repositories/accounts/accounts.module';
+import { RecordsModule } from '../../records/records.module';
+import { BudgetsModule } from '../../budgets/budgets.module';
+import { BudgetHistoryModule } from '../../budget-history/budget-history.module';
 import { AccountsActionsService } from './services/accounts.-actions.service';
 
 @Module({
