+++ src/repositories/accounts/accounts.interface.ts
@@ -1,5 +1,5 @@
 import { Types } from 'mongoose';
-import { GeneralResponse } from '../response.interface';
+import { GeneralResponse } from '../../response.interface';
 import { Account as AccountEntity } from './entities/accounts.entity';
 
 export interface AccountModel extends AccountEntity {
